backend/requirements.txt
django
django-environ
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
gunicorn
requests
pillow
django-crontab
channels
daphne
channels_redis
